refactor: step1 피드백사항 리펙터링
- 벨리드 메서드 위치는 도메인에 맞게 변경
- stream 액션별 개행 변경
- 외부에서 사용하지 않는 클래스 접근제어자 수정
- 단위테스트에 @SpringBootTest 제거
- 단위테스트 displayName 수정
